Limpiado código, nueva estructura. Funcionalidad de actualizar y borrar anuncio básica.
