--- app/src/main/java/com/hyenatechnologies/connections/Data.java
@@ -1,17 +1,0 @@
-package com.hyenatechnologies.connections;
-
-// Data class
-public class Data {
-    public void getAnuncio(final String id) {
-    }
-
-    public void getAnimal(final String id) {
-    }
-
-    public void getInfo(final String id) {
-    }
-
-    public void setData(final String id, final String op) {
-    }
-
-}
