--- app/src/main/java/com/hyenatechnologies/connections/Logger.java
@@ -1,10 +1,0 @@
-package com.hyenatechnologies.connections;
-
-import java.util.List;
-
-public class Logger {
-
-    public void doLogin(final List<String> data) {
-    }
-
-}
