--- app/src/main/java/com/hyenatechnologies/connections/Register.java
@@ -1,11 +1,0 @@
-package com.hyenatechnologies.connections;
-
-import java.util.List;
-
-
-public class Register {
-
-    public void doRegister(final List<String> data) {
-    }
-
-}
