--- app/src/main/java/com/hyenatechnologies/wallapet/Cons.java
@@ -1,10 +1,0 @@
-package com.hyenatechnologies.wallapet;
-
-/**
-* Clase que contiene constantes.
- */
-public class Cons {
-
-    public static final int SUCCESS = 1;
-    public static final int ERROR = 1;
-}
