+++ app/src/main/java/com/hyenatechnologies/wallapet/pantallas/Login.java
@@ -1,24 +1,26 @@
-package com.hyenatechnologies.wallapet;
+package com.hyenatechnologies.wallapet.pantallas;
 
 import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
 import android.view.Menu;
 import android.view.MenuItem;
 
+import com.hyenatechnologies.wallapet.R;
 
-public class MainActivity extends ActionBarActivity {
+
+public class Login extends ActionBarActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_pantalla_principal);
+        setContentView(R.layout.activity_login);
     }
 
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main, menu);
+        getMenuInflater().inflate(R.menu.menu_login, menu);
         return true;
     }
