+++ app/src/main/java/com/hyenatechnologies/wallapet/pantallas/PantallaPrincipal.java
@@ -1,4 +1,4 @@
-package com.hyenatechnologies.wallapet;
+package com.hyenatechnologies.wallapet.pantallas;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,8 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+
+import com.hyenatechnologies.wallapet.R;
 
 /**
  * WallaPet Android App
@@ -60,7 +62,7 @@ public class PantallaPrincipal extends ActionBarActivity {
             @Override
             public void onClick(View v) {
                 //Lanzamos la actividad de crear
-                Intent i = new Intent(getApplicationContext(), CrearAnuncio.class);
+                Intent i = new Intent(getApplicationContext(), CrearModificarAnuncio.class);
 
                 startActivity(i);
