+++ app/src/main/res/layout/activity_crear_anuncio.xml
@@ -5,7 +5,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:orientation="vertical"
-    tools:context="com.hyenatechnologies.wallapet.CrearAnuncio">
+    tools:context="com.hyenatechnologies.wallapet.pantallas.CrearModificarAnuncio">
 
 
     <EditText
