+++ app/src/main/res/layout/activity_vista_anuncio.xml
@@ -97,6 +97,20 @@
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:text="Especie: "
+            />
+        <TextView
+            android:id="@+id/anuncioEspecie"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content" />
+    </LinearLayout>
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
             android:text="Precio: "
             />
         <TextView
@@ -104,5 +118,17 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content" />
     </LinearLayout>
+
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Borrar anuncio"
+        android:id="@+id/botonBorrar" />
+
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Actualizar anuncio"
+        android:id="@+id/botonActualizar" />
 
 </LinearLayout>
