+++ app/src/main/res/menu/menu_registro.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.com.hyenatechnologies.wallapet.Registro">
+    tools:context="com.com.hyenatechnologies.wallapet.pantallas.Registro">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" app:showAsAction="never" />
 </menu>
